+++ src/webgl/p5.RendererGL.js
@@ -5,7 +5,7 @@ import libtess from 'libtess'; // Fixed with exporting module from libtess
 import './p5.Shader';
 import './p5.Camera';
 import Renderer from '../core/p5.Renderer';
-import './p5.Matrix';
+import '../math/p5.Matrix';
 import './p5.Framebuffer';
 import { MipmapTexture } from './p5.Texture';
